Mejoras en gestión de audio y temporizador del juego
Se han añadido nuevas clases como `AudioManager`, `GameTimer` y `GameStats` para gestionar la música ambiental y el temporizador, mejorando la estructura del código. Se han eliminado las referencias a `SoundPlayer` en las clases de nivel, simplificando la gestión del audio.

Además, se ha centralizado la inicialización de componentes en métodos específicos, se han añadido métodos para verificar la completitud del Sudoku y resaltar conflictos, y se han realizado mejoras en la interfaz de usuario. Se ha optimizado el código eliminando comentarios y código redundante.
